requirements.txt
Flask
sounddevice
soundfile
numpy
requests
RPi.GPIO
gtts
openai
openai-whisper